Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, Text, FlatList } from 'react-native';
+import { View, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { LEADERS } from '../shared/leaders';
 import { connect } from 'react-redux';
@@ -33,7 +33,24 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id);
+                    onPress: () => {
+                        Alert.alert(
+                          'Delete Favorite?'
+                          'Are you sure you wish to delete the favorite dish' + item.name + '?',
+                            [
+                                { text: 'Cancel',
+                                   onPress: () => console.log(item.name + 'Not Deleted')
+                                   style: 'cancel'
+                                },
+                                {
+                                    text: 'OK'
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                           ],
+                           { cancelable: false }
+                        );
+                    }
+                    
                 }
             ];
             return(
